src/services/trash-detection/requirements.txt
fastapi==0.115.5
Pillow==11.0.0
Requests==2.32.3
torch==2.5.1
uvicorn==0.32.0
timm==1.0.11
python-dotenv=1.0.1
